desktop dialog implemented
