--- composeApp/src/androidMain/kotlin/ua/smartwaste/kmp/presentation/screens/bucket/AddRubbishSheet.kt
@@ -1,21 +1,0 @@
-@file:OptIn(ExperimentalMaterialApi::class, ExperimentalMaterial3Api::class)
-
-package ua.smartwaste.kmp.presentation.screens.bucket
-
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-
-/**
- * Created by gle.bushkaa email(gleb.mokryy@gmail.com) on 12/27/2023
- */
-
-@Composable
-fun AddRubbishSheet(
-    modifier: Modifier = Modifier,
-    dismissRequest: () -> Unit,
-) {
-}
