+++ composeApp/src/commonMain/kotlin/ua/smartwaste/kmp/presentation/components/SmartOutlinedIconButton.kt
@@ -3,8 +3,8 @@ package ua.smartwaste.kmp.presentation.components
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Icon
-import androidx.compose.material.OutlinedButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.OutlinedButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
